+++ src/ActionGameCore/Assets/CharacterControls/CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added a parameter to make the Pull-up module operable on the ground.
 ### Fixed
 - Fixed a bug where it would hop on slopes.
+- Fixed friction calculation.
 
 ## [0.10.1] - 2024-08-08
 ### Changed